+++ src/lib/video_examiner.h
@@ -32,10 +32,10 @@ class VideoExaminer
 {
 public:
 	virtual ~VideoExaminer () {}
-	virtual boost::optional<float> video_frame_rate () const = 0;
+	virtual boost::optional<double> video_frame_rate () const = 0;
 	virtual dcp::Size video_size () const = 0;
 	virtual Frame video_length () const = 0;
-	virtual boost::optional<float> sample_aspect_ratio () const {
-		return boost::optional<float> ();
+	virtual boost::optional<double> sample_aspect_ratio () const {
+		return boost::optional<double> ();
 	}
 };
